requirements.txt
PyYAML==3.11
requests==2.7.0
schema==0.3.1
